Fixed issues with caching strongswan and installing the IPOP deb
Resolved the issues with checking out and caching strongswan by
doing such in a temp directory and then tossing the result into
a tarball, which is then copied to a temp directory, untarred,
compiled and installed when the VM boots. Next step is to cache
the compiled version. A similar trick is applied to the deb
package, which strangely errors when attempting to copy or to
install it from the shared folder.