Now cache compiled version of strongSwan and loop IPOP install
We now run make before tarring the strongSwan source directory.
Due to repeated "protocol errors" when attempting to copy the
IPOP deb package to a temporary directory, in addition to tarring
the file, we repeat the IPOP installation portion until success
is had. Overall, the time for first boot of a VM has been reduced
by about half.